pyproject.toml
[project]
name = "SpeedPortAPI"
description = ""
version = "0.1.0"
requires-python = ">=3.12"
readme = "README.md"
authors = [{ name = "Ozkar L. Garcell", email = "ozkar.garcell@gmail.com" }]
dependencies = [
    "fastapi[standard]>=0.115.3",
    "hishel>=0.0.33",
    "httpx>=0.27.2",
    "orjson>=3.10.10",
    "pycryptodome>=3.21.0",
    "pydantic-extra-types>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pyhstr>=0.5.2",
    "python-dotenv>=1.0.1",
]
[tool.uv]
dev-dependencies = [
    "pre-commit>=4.0.1",
    "ruff>=0.7.0",
    "ruff-lsp>=0.0.58",
    "uvicorn>=0.32.0",
    "pytest>=8.3.3",
    "pytest-cov>=6.0.0",
    "respx>=0.21.1",
]

[tool.pytest.ini_options]
addopts = "-vv --cov=app --cov-report=xml"
